requirements.txt
azure-core==1.9.0
azure-cosmos==4.2.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flatten-dict==0.3.0
freeze==3.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pathlib2==2.3.5
requests==2.25.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
